% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object.R
\name{get_object}
\alias{get_object}
\title{Get something out of the cache}
\usage{
get_object(object_name)
}
\arguments{
\item{object_name}{Name of what to get}
}
\value{
thing
}
\description{
Get something out of the cache
}
